+++ .changeset/perfect-cobras-do.md
@@ -1,0 +1,5 @@
+---
+"app-avatax": patch
+---
+
+Debugging: Disable importing dynamodb in checkout_calculate_taxes. This will disable logs feature even if flag is enabled
