+++ apps/avatax/src/pages/api/webhooks/checkout-calculate-taxes.ts
@@ -11,10 +11,8 @@ import { SubscriptionPayloadErrorChecker } from "@/lib/error-utils";
 import { createLogger } from "@/logger";
 import { loggerContext } from "@/logger-context";
 import { CalculateTaxesUseCase } from "@/modules/calculate-taxes/use-case/calculate-taxes.use-case";
-import { clientLogsFeatureConfig } from "@/modules/client-logs/client-logs-feature-config";
-import { DynamoDbLogWriter, ILogWriter, NoopLogWriter } from "@/modules/client-logs/log-writer";
-import { LogWriterFactory } from "@/modules/client-logs/log-writer-factory";
-import { LogsRepositoryDynamodb } from "@/modules/client-logs/logs-repository";
+import { ILogWriter, LogWriterContext } from "@/modules/client-logs/log-writer";
+import { NoopLogWriter } from "@/modules/client-logs/noop-log-writer";
 import { AvataxInvalidAddressError } from "@/modules/taxes/tax-error";
 import { checkoutCalculateTaxesSyncWebhook } from "@/modules/webhooks/definitions/checkout-calculate-taxes";
 
@@ -31,7 +29,12 @@ const withMetadataCache = wrapWithMetadataCache(metadataCache);
 const subscriptionErrorChecker = new SubscriptionPayloadErrorChecker(logger, captureException);
 const useCase = new CalculateTaxesUseCase({
   configExtractor: new AppConfigExtractor(),
-  logWriterFactory: new LogWriterFactory(),
+  logWriterFactory: {
+    createWriter(context: LogWriterContext): ILogWriter {
+      // temp - until we fix
+      return new NoopLogWriter();
+    },
+  },
 });
 
 /**
